+++ config.yml
@@ -8,18 +8,18 @@ aml_experiment_details:
 
 performance_evaluation:
   inputs:
-    datastore: "bias_analysis_old"
-    ground_truth_rel_path: "validation-to-tag"
-    predictions_rel_path: "inference/yolov8"
+    datastore: "bias_analysis_sampled_panos"
+    ground_truth_rel_path: "kavel10_kempkes_geo360_sampled_panos"
+    predictions_rel_path: "inference/yolov5"
   outputs:
-    datastore: "bias_analysis_old"
-    output_rel_path: "evaluation/v2-2/v2.2_test2"
+    datastore: "bias_analysis_sampled_panos"
+    output_rel_path: "evaluation/v5/test"
   categories_json_path: "categories_eval.json"
-  mapping_json_path: "groupings_old_bias_analysis.json"
-  dataset_name: "v2-2"
-  model_name: "yolov8m_1280_v2.2"
+  mapping_json_path: "groupings_wheelchairs_bias_analysis.json"
+  dataset_name: "sampled_panos"
+  model_name: "yolov5"
   ground_truth_image_shape: [3840, 2160]
-  predictions_image_shape: [1280, 720]
+  predictions_image_shape: [2048, 1024]
   prediction_labels_rel_path: "labels"
   splits: []
   target_classes: []
